% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-read-write.R
\name{clean_struct}
\alias{clean_struct}
\title{Clean the imported matlab .mat matrix}
\usage{
clean_struct(x)
}
\arguments{
\item{x}{an imported *_results.mat file}
}
\description{
Clean the imported matlab .mat matrix
}
