% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/100-install.R
\name{install_octave}
\alias{install_octave}
\title{Install Octave}
\usage{
install_octave(
  keep_install_file = FALSE,
  download_dir = tempdir(),
  silent = FALSE,
  ...
)
}
\description{
Install Octave
}
