% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-read-write.R
\name{mat_write}
\alias{mat_write}
\title{Reads a MAT file}
\usage{
mat_write(matfile, ...)
}
\arguments{
\item{matfile}{a MAT file}

\item{...}{further arguments passe to \code{\link[R.matlab]{readMat}}.}
}
\description{
This function is a soft-wrapper around \code{\link[R.matlab]{readMat}}.
}
