% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-run.R
\name{oct}
\alias{oct}
\title{Run Octave}
\usage{
oct(
  code,
  args = character(),
  spinner = FALSE,
  echo_cmd = FALSE,
  echo = FALSE,
  timeout = Inf,
  wd = NULL,
  ...
)
}
\description{
Run Octave
}
\examples{
\dontrun{
oct("2+2")
}
}
