% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\name{OctaveREPL}
\alias{OctaveREPL}
\title{Repl octave}
\description{
Repl octave

Repl octave
}
\section{Super class}{
\code{\link[octaver:OctaveSession]{octaver::OctaveSession}} -> \code{OctaveRepl}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OctaveREPL$new()}}
\item \href{#method-prompt}{\code{OctaveREPL$prompt()}}
\item \href{#method-clone}{\code{OctaveREPL$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="octaver" data-topic="OctaveSession" data-id="eval">}\href{../../octaver/html/OctaveSession.html#method-eval}{\code{octaver::OctaveSession$eval()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="octaver" data-topic="OctaveSession" data-id="finalize">}\href{../../octaver/html/OctaveSession.html#method-finalize}{\code{octaver::OctaveSession$finalize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="octaver" data-topic="OctaveSession" data-id="kill">}\href{../../octaver/html/OctaveSession.html#method-kill}{\code{octaver::OctaveSession$kill()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="octaver" data-topic="OctaveSession" data-id="state">}\href{../../octaver/html/OctaveSession.html#method-state}{\code{octaver::OctaveSession$state()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OctaveREPL$new(bin = find_octave(), params = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prompt"></a>}}
\if{latex}{\out{\hypertarget{method-prompt}{}}}
\subsection{Method \code{prompt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OctaveREPL$prompt(prompt)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OctaveREPL$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
